backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
redis[hiredis]
aioredis
authlib
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aio-pika
pydantic
pydantic-settings
email-validator
Pillow
python-slugify
websockets
starlette
itsdangerous
pytest
pytest-asyncio
httpx
anyio
